Move `mephenger/models` -> `mephenger/legacy/models`
