+++ mephenger/views/channels.py
@@ -15,7 +15,7 @@ from kivy.uix.screenmanager import ScreenManagerException
 from kivy.uix.scrollview import ScrollView
 
 from mephenger.config import config
-from mephenger.models.screens_manager import ScreensManager
+from mephenger.legacy.models.screens_manager import ScreensManager
 
 
 Builder.load_file("{0}/channel.kv".format(config.VIEWS_DIR))
